Refactor provider setup into thematic groups
